add user center.

+++ src/main/java/com/zuoxiaolong/servlet/UploadImage.java
@@ -1,19 +1,22 @@
 package com.zuoxiaolong.servlet;
 
-import com.zuoxiaolong.config.Configuration;
-import com.zuoxiaolong.dao.UserDao;
-import com.zuoxiaolong.util.IOUtil;
-import com.zuoxiaolong.util.ImageUtil;
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
+import org.apache.commons.lang.StringUtils;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import java.io.File;
-import java.io.IOException;
-import java.util.List;
+import com.zuoxiaolong.config.Configuration;
+import com.zuoxiaolong.dao.UserDao;
+import com.zuoxiaolong.util.IOUtil;
+import com.zuoxiaolong.util.ImageUtil;
 
 /*
  * Copyright 2002-2015 the original author or authors.
@@ -46,20 +49,22 @@ public class UploadImage extends AbstractServlet {
 		String path = null;
 		DiskFileItemFactory factory = new DiskFileItemFactory(5 * 1024, new File(Configuration.getContextPath("temp")));
 		ServletFileUpload upload = new ServletFileUpload(factory);
-		upload.setSizeMax(1024 * 1024);
+		upload.setSizeMax(3 * 1024 * 1024);
 		try {
 			List<FileItem> items = upload.parseRequest(request);
-			if (items != null && items.size() > 0) {
-				FileItem fileItem = items.get(0);
+			FileItem fileItem = null;
+			if (items != null && items.size() > 0 && StringUtils.isNotBlank((fileItem = items.get(0)).getName())) {
 				path = ImageUtil.generatePath(fileItem.getName());
 				IOUtil.copy(fileItem.getInputStream(), Configuration.getContextPath(path));
 				fileItem.delete();
+				String url = Configuration.getSiteUrl(path);
+				if (UserDao.uploadImage(getUsername(), url)) {
+					writeText(url);
+				}
 			}
 		} catch (FileUploadException e) {
 			throw new RuntimeException(e);
 		}
-		UserDao.uploadImage(getUsername(), Configuration.getSiteUrl(path));
-		getResponse().sendRedirect(Configuration.getSiteUrl("/common/user_center.ftl"));
 	}
 
 }
